+++ package.json
@@ -16,6 +16,7 @@
     "@types/node": "^20.8.6",
     "@types/react": "^18.2.28",
     "@types/react-dom": "^18.2.13",
+    "@vercel/postgres-kysely": "^0.5.0",
     "antd": "^5.10.1",
     "autoprefixer": "^10.4.16",
     "eslint": "^8.51.0",
@@ -23,11 +24,10 @@
     "jsonwebtoken": "^9.0.2",
     "kysely": "^0.26.3",
     "kysely-codegen": "^0.11.0",
-    "kysely-planetscale": "^1.4.0",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
-    "mysql2": "^3.6.2",
     "next": "^13.5.5",
+    "pg": "^8.11.3",
     "postcss": "^8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
