+++ src/app/actions/kysely.ts
@@ -1,11 +1,4 @@
-import { Kysely } from 'kysely';
+import { createKysely } from '@vercel/postgres-kysely';
 import { DB } from 'kysely-codegen';
-import { PlanetScaleDialect } from 'kysely-planetscale';
 
-export const kysely = new Kysely<DB>({
-  dialect: new PlanetScaleDialect({
-    host: process.env.DB_HOST,
-    username: process.env.DB_USERNAME,
-    password: process.env.DB_PASSWORD,
-  }),
-});
+export const kysely = createKysely<DB>();
